WL-0MM2FA7GN10FQZ2R: Extract fallback search tests into tests/search-fallback.test.ts
